include/airflow_task_runner/requirements.txt
virtualenv==20.21.0
jinja2==3.1.2
snowflake-snowpark-python[pandas]==1.5.1 
snowflake-ml-python==1.0.2 
prometheus-fastapi-instrumentator==6.0.0